requirements.txt
openpyxl==3.0.7
Pillow==9.1.0
PyExifTool==0.4.11
